Revert "Use CSetOffset for stack-pointer to stack-capability conversions so that we don’t hit edge cases with the base of the stack."
For now we'll just hope no one tries to allocate something right at the
bottom of the stack.  This caused issues with setting the length,
because the offset would be long past the end with the new length.  On
CHERI128, this is not a problem, because CSetBound would correctly set
the length.

This reverts commit 9b0ffa07c0604114e27cc0d9ca41aae38b822c54.