Add a pass that turns PCC-relative calls into PC-relative calls.
This is needed because the inliner does not understand that a call of a PCC-relative offset is a valid inlining target.

We don’t need a converse transform, because if you are using the pure-capability ABI and try to do a direct call, the back end automatically does the PCC derivation (and, currently, the CHERI pipeline is such that it’s safe to use the register value immediately after defining it, so there are no instruction scheduler implications from doing the expansion late).

We do need this pass, because clang’s function-to-pointer decay logic is invoked for both taking the address of a function to call and taking the address of a function to use as a function pointer.