Fix capsizefix base and length fixup when we fall back to using the section bounds
When we are not able to find the size of a symbol, we conservatively use the section
bounds. The bug causes the base to be set to the symbol address and the length to
the section size, leading to capabilities that go outside the section bounds.
The fix sets base and length to be equal to the section start and size and the
capability offset is set to the symbol address.

Fix minor bug in llvm-objdump -C reporting incorrect capability relocation permissions.