MIPS N32 patch 1, exposes -mabi flag
Replacing the hidden -target-abi flag, so all the .ll and .s changes
just replace RUN: ... -target-abi with RUN: -mabi.

From https://reviews.llvm.org/D21465 :

[llc+llvm-mc] Replace the hidden -target-abi option with a -mabi thats
visible in --help.

Although this patch is a cosmetic change, it is important for MIPS since
it
is a prerequisite of moving the ABI to the environment component of the
triple (much like X86's and ARM's GNUX32, GNUEABI, and GNUEABIHF) which
in
turn is important for both enabling IAS by default for the N64 ABI and
for
distinguishing between the N32 and N64 in order to fix IAS for N32 and
prevent inappropriate interlinking in the IRLinker. Follow-up patches
will try
to modify the triple according to -mabi for targets that prefer this
instead of
passing it to MCTargetOptions::ABIName.

See http://reviews.llvm.org/D20916 (which this patch has been extracted
from) for more detail on the wider context of this patch as well as an
explanation as to why having the ABI name in MCTargetOptions does not
work
for Mips.

The test changes are all switching -target-abi to -mabi.

A similar patch will be posted for 'clang -cc1' and 'clang -cc1as' (the
latter currently accepts a -target-abi option but ignores it, I'll fix
this
at the same time).

Patch By: Daniel Sanders