Ignore undef arguments when lowering IR calls
Clang inserts undef argument values to add padding before a vararg union/struct for alignment purposes. For example:

call void (i32, ...) @a(i32 signext 99, i64 undef, i8 addrspace(200)* inreg %1)

These undef values are not taken into account when initailsing capability
register C13 (used for passing varargs in the pure capability ABI).  In
particular, C13's base is set to the undef value's address rather than the
struct even if the struct is on an alignment boundary (as a result of preceding
arguments being passed in registers). This leads to the callee not being able
to load the vararg union correctly.

Fixes #232.