Don't generate a redundant CIncOffset when spilling.
If we're spilling or restoring at an offset that's larger than the
immediate field in a capability load / store, we materialise a large
immediate in an integer register.  If the load / store instruction is
using $zero as its integer offset register, then we can use the new
register directly, rather than computing a new offset.

Note: This still generates some quite suboptimal assembly, because
we're materialising lots of constants when we have lots of large spills
and we could be reusing some of them.  The LLVM back end isn't really
set up for doing this, because the assumption is that only really huge
functions require spills that don't fit in immediates.