Initial support for selectable capability sizes.
With these changes, it should be possible to use a single LLVM back end for all capability sizes, rather than having to compile different toolchains for CHERI256 and CHERI128.

Currently, one LLVM test is failing, and no executable tests have been run, so this should be regarded as highly experimental!