Fix remaining failing tests.
Khilan's cheri-varargs-undef.ll is marked XFAIL.  This test is fragile
and seems to break at the slightest change in code generation.  It
doesn't document what it's actually trying to test, so it's impossible
to tell whether it is failing because the thing that it was testing is
broken, or because the test makes too many assumptions about code
generation.