When disassembling add a comment which symbol is loaded from the captable
This only works for the single instruction case and assumes that every load
relative to $c26 is a cap table load but for most cases this should be
useful when looking at the disassembly.
I can't handle the mxcaptable case yet because I don't know if there is a
way to infer the value of the register operand for a load.