Rework memory operation lowering.
We now no longer expand memset, memcpy and memmove intrinsics in an IR
pass prior to code generation for non-capability arguments.  Instead,
during DAG construction, we use the provided hook to emit a call to the
relevant function (or the _c suffixed version in the hybrid ABI).

This change also slightly refines the expansion of some of these
sequences into inline instructions.  In particular, we should now be
emitting capability stores for long zero stores.

Note that we currently only emit capabilities for bzero when the size
is a multiple of capability size because SelectionDAG doesn't pad with
smaller stores, it emits overlapping stores (which won't work with
capabilities).