Improve bzero generation.
If we have something small that is aligned on a capability boundary, but
not a multiple of capability sized, then generate and efficient zeroing
code using capability stores for most of it and then 1-4 other
instructions for the rest.  In most cases, we need only two other
instructions.

The CHERI loads and stores don't allow unaligned immediate offsets, so
if we want to do an unaligned store to write at the end then we need an
extra instruction to generate the address.  We currently do this only
if we can't write the last 1-7 bytes in any other way, because two
adjacent stores with independent bases are likely to be faster (or, at
least, no slower) than a store that depends on the result of the
previous address calculation instruction.