Remove some dead and broken code.
The addressing mode folder tried to hoist GOT loads out of loops, but
that's no longer required and the code doing it was also moving other
loads and stores out of loops.  Moving stores broke dependencies (it
moved them to before the register containing the result was written)
and so caused the compiler to crash, moving loads just generated
incorrect code.

Fixes #245