Fix lowering of offsets from frame index.
Most stack spills and reloads are at offset 0 from a frame index.
Spills and reloads of `%lo` and `%hi` put both in the same frame index,
but then use an offset within the frame index.  This is the only case
where the offset is used, so the fact that we were ignoring it was only
apparent in functions that perform a division or multiplication before
a function call and don't use the result until afterwards (and so must
spill the accumulator registers, rather than registers that they have
moved the accumulator to for use by other instructions).

Fixes #258