Add a --cheri-tests-filter flag to lit to allow running only CHERI tests
We currently run the full test suite for both CHERI128 and CHERI256 with
the only difference being different expansions of %cheri_foo commands.
This wastes lots of time rerunning tests that will not change behaviour.
This flag adds features to UNSUPPORTED/REQUIRED in order to either include,
exclude or only run CHERI tests.

I've also added  check-all-cheri{128,256}-only ninja targets that can be
used as a shortcut so I don't have to remember the name of the flag.

We currently still run the unit tests twice since this only filters the
shtest test format but I will fix that in a follow-up commit