Update MIPS asm parser to handle clwc1/cswc1+cldc1/csdc1 again
It was doing
`getReg(is64Bit ? Mips::GPR64RegClassID : Mips::GPR32RegClassID, getATReg(IDLoc));`
which would get the real at reg number and then look up that as an index in
the other register class. This happened to work before but now the register
numbers have changed so it gives a nonsense value instead. Fixed by just
using getATReg() and returning an error if $at is not available