Revert "Make always_inline work at -O0 with CHERI cap-table"
Unfortunately it appears this broke /sbin/init in CheriBSD. While this
does fix a real issue, it goes away at -O1 or higher whereas there is no
known workaround for the CheriBSD boot issue. It is also really hard to
get a reduced testcase since the miscompilation could be almost anywhere
in the CheriBSD sources.

This reverts commit a44e103c515964c23f27b7ecc0402733105aca0d.