Log the CSetBounds generated by the CheriPureCapABI pass
This sets bounds on all stack variables. Also did a bit of refactoring so
that the individual calls to add() don't have to manually infer the source
location from debug instructions. It requires some annoying special casing
of AllocaInst* and therefore should rather be done in a shared function.
AllocaInst don't have the DebugLoc attached directly, instead there is a
@llvm.dbg.declare() call later on that needs to be parsed instead.